requirements.txt
flask
python-dotenv
stripe
gunicorn
